userRegistration/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pillow==10.2.0
psycopg==3.1.18
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
twilio==8.13.0
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
